Revert "Floating point probabilities"
This reverts commit ac1800fa773a22c14f04e7364bdb0e074326419d.

+++ otherblocks/src/com/sargant/bukkit/otherblocks/OtherBlocksContainer.java
@@ -9,6 +9,6 @@ public class OtherBlocksContainer
 	public Material tool;
 	public Integer quantity;
 	public Integer damage;
-	public Double chance;
+	public Integer chance;
 	public short color;
 }
